// src/pages/mypage/types/review.ts
import type { Category } from "./catetory";

export interface Review {
  id: number;
  restaurantId?: number;
  restaurantName: string;
  content: string;
  rating: number; // 0 ~ 5
  likeCount: number;
  commentCount: number;
  liked: boolean;
  pinned: boolean;
  createdAt: string;
  updatedAt?: string;
  avgRating: number; // 평균 평점
  reviewCount: number; // 리뷰 개수
  categories: Category[];
  address: string;
  imageUrl?: string | null;
}

export interface ReviewPage {
  reviews: Review[];
  nextCursor?: number;
  prevCursor?: number;
}